refactor set-job
split set-job out into seprarate functions - 
set-jobowner *includes check that account running has sysadmin permissions on server)
set-jobnotifications (if job has an operator included in xml then set the notifications to operator)
set-jobproperties (orchestrates the job notifcations and job owner funcrtions and sets the description/category/enabled status/target server)
set-job (creates the job if does not exist. If does then gets from server. Then calls set-jobproperties)

Idea for this is to make it easier to debug/spot where an issue is and to be more flexible with permissions